Move ga-action toggle to separate file
This allows us to simplify header.js
