+++ assets/src/js/header.js
@@ -2,25 +2,8 @@
 
 import {setupAccessibleNavMenu, updateNavMenuTabIndex} from './header/accessibleNavMenu';
 import setupMobileTabsMenuScroll from './header/setupMobileTabsMenuScroll';
+import {setSearchToggles, toggleGaActionAttributes} from './header/setupGaActions';
 import {setupCloseNavMenuButton, setupDocumentClick, lockScrollWhenNavMenuOpen} from './header/setupNavMenu';
-
-const updateGaAction = (element, elementName) => {
-  element.dataset.gaAction = `${element.getAttribute('aria-expanded') === 'false' ? 'Open' : 'Close'} ${elementName}`;
-};
-
-/**
- * Propagate attributes to all search toggles
- *
- * @param {boolean} expanded Toggle is expanded
- */
-const setSearchToggles = expanded => {
-  const toggles = document.querySelectorAll('.nav-search-toggle');
-  toggles.forEach(toggle => {
-    toggle.setAttribute('aria-expanded', expanded ? 'true' : 'false');
-    updateGaAction(toggle, 'search');
-    toggle.classList.toggle('open', expanded);
-  });
-};
 
 const toggleNavElement = element => {
   const target = element.dataset.bsTarget;
@@ -63,21 +46,7 @@ const toggleNavElement = element => {
   updateNavMenuTabIndex();
 
   // Toggle data-ga-action attribute used in GTM tracking.
-  const countryDropdownToggle = document.querySelector('.country-dropdown-toggle');
-  const countrySelectorToggle = document.querySelector('.country-selector-toggle');
-  const navMenuToggle = document.querySelector('.nav-menu-toggle');
-
-  if (countryDropdownToggle) {
-    updateGaAction(countryDropdownToggle, 'Country Selector');
-  }
-
-  if (countrySelectorToggle) {
-    updateGaAction(countrySelectorToggle, 'Country Selector');
-  }
-
-  if (navMenuToggle) {
-    updateGaAction(navMenuToggle, 'Menu');
-  }
+  toggleGaActionAttributes();
 };
 
 export const setupHeader = () => {
